@@ -16,12 +16,15 @@
     "@babel/plugin-transform-arrow-functions": "^7.12.1",
     "@babel/plugin-transform-block-scoping": "^7.12.12",
     "@growthbunker/vueflags": "^0.1.2",
+    "@vue/compat": "^3.1.1",
+    "@vue/compiler-sfc": "^3.1.1",
     "autoprefixer": "^10.2.6",
     "axios": "^0.21.1",
     "bootstrap": "^5.0.1",
     "core-js": "^3.7.0",
     "es6-promise": "^4.1.1",
     "express": "^4.17.1",
+    "hammerjs": "^2.0.8",
     "i18n-iso-countries": "^6.0.0",
     "lodash": "^4.17.4",
     "luxon": "^1.25.0",
@@ -32,16 +35,15 @@
     "request": "^2.88.2",
     "sass": "^1.34.1",
     "sass-loader": "^12.0.0",
-    "vue": "^2.5.16",
-    "vue-cookie": "^1.1.4",
-    "vue-gtag": "^1.7.0",
-    "vue-i18n": "^8.22.0",
-    "vue-nl2br": "^0.1.2",
-    "vue-router": "^3.3.3",
+    "vue": "^3.1.0",
+    "vue-gtag": "^2.0.1",
+    "vue-i18n": "^9.1.6",
+    "vue-router": "4",
     "vue-select": "^3.10.3",
-    "vue-toast-notification": "^0.5.4",
-    "vue-touch": "^2.0.0-beta.4",
-    "vuex": "^3.3.0"
+    "vue-toast-notification": "https://github.com/ankurk91/vue-toast-notification.git#next",
+    "vue3-nl2br": "^0.1.5",
+    "vue3-touch-events": "^4.1.0",
+    "vuex": "^4.0.1"
   },
   "devDependencies": {
     "@babel/core": "^7.12.3",
@@ -65,12 +67,8 @@
     "semver": "^5.4.1",
     "shelljs": "^0.7.8",
     "style-loader": "^2.0.0",
-    "terser-webpack-plugin": "^5.0.3",
     "url-loader": "^4.1.1",
-    "vue-loader": "^15.9.4",
-    "vue-template-compiler": "^2.6.12",
-    "vue-toastification": "^1.7.8",
-    "vue2-filters": "^0.3.0",
+    "vue-loader": "^16.0.0",
     "webpack": "^5.3.1",
     "webpack-cli": "^4.1.0"
   },
