@@ -13,7 +13,7 @@
          v-bind:class="{ 'program-current': isCurrent, 'long-enough': isLongEnough }">
       <div class="program-inner" v-bind:title="title">
         <div class="program-img" v-if="program.picture_url && (hover || isCurrent)" v-once>
-          <img v-bind:src="program.picture_url | picture" alt="" @mousedown.prevent="">
+          <img v-bind:src="picturePath" alt="" @mousedown.prevent="">
         </div>
         <div class="program-infos" :style="infosStyle">
           <div class="program-title" v-once>
@@ -29,7 +29,7 @@
                 {{ currentSong }}
               </span>
               <span class="program-description-short-inner-text" v-once>
-                {{ program.description | shorten(program.duration) }}
+                {{ shorten(program.description, program.duration) }}
               </span>
             </div>
           </div>
@@ -49,6 +49,9 @@ import { TIMEZONE, THUMBNAIL_PROGRAM_PATH, PROGRAM_LONG_ENOUGH } from '../../con
 import ScheduleRadioSection from './ScheduleRadioSection.vue';
 
 export default {
+  compatConfig: {
+    MODE: 3
+  },
   components: { ScheduleRadioSection },
   props: ['program', 'radioPlaying'],
   data() {
@@ -124,13 +127,11 @@ export default {
       return this.program.end_overflow;
       // return this.$store.state.schedule.scheduleDisplay[this.program.hash].container.nextDayOverflow;
     },
-  },
-  methods: {
-    detailClick() {
-      // this.displayDetail = !this.displayDetail;
+    picturePath() {
+      return `${THUMBNAIL_PROGRAM_PATH}${this.program.picture_url}`;
     }
   },
-  filters: {
+  methods: {
     shorten(value, duration) {
       if (value === null) {
         return '';
@@ -141,7 +142,9 @@ export default {
 
       return value.split('\n')[0];
     },
-    picture: value => `${THUMBNAIL_PROGRAM_PATH}${value}`
-  },
+    detailClick() {
+      // this.displayDetail = !this.displayDetail;
+    }
+  }
 };
 </script>
