@@ -29,9 +29,12 @@ import { mapGetters, mapState } from 'vuex';
 import StreamsListFilters from './StreamsListFilters.vue';
 import StreamsListPagination from './StreamsListPagination.vue';
 import StreamsListOne from './StreamsListOne.vue';
-import Loading from '../Loading.vue';
+import Loading from '../Utils/Loading.vue';
 
 export default {
+  compatConfig: {
+    MODE: 3
+  },
   components: {
     StreamsListFilters,
     StreamsListPagination,
