@@ -1,11 +1,10 @@
-import Vue from 'vue';
-import VueI18n from 'vue-i18n';
+import { createI18n } from 'vue-i18n';
 import { messages, dateTimeFormats } from './lang';
 
-Vue.use(VueI18n);
-
 /* eslint-disable no-undef */
-const i18n = new VueI18n({
+const i18n = createI18n({
+  // legacy: false,
+  // globalInjection: true,
   fallbackLocale: 'fr',
   locale,
   messages,
